J.A.R.V.I.S ~ Deep Learning/backend/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from transformers import AutoModelForCausalLM, AutoTokenizer
import torch
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="J.A.R.V.I.S API")

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000"],  # Frontend URL
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Initialize model and tokenizer
model_name = "microsoft/DialoGPT-medium"  # We'll use this as a starting point
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForCausalLM.from_pretrained(model_name)

class UserInput(BaseModel):
    message: str

@app.post("/api/chat")
async def chat_with_jarvis(user_input: UserInput):
    try:
        # Encode the input text
        input_ids = tokenizer.encode(user_input.message + tokenizer.eos_token, return_tensors="pt")
        
        # Generate response
        with torch.no_grad():
            response_ids = model.generate(
                input_ids,
                max_length=1000,
                pad_token_id=tokenizer.eos_token_id,
                temperature=0.7,
                num_return_sequences=1
            )
        
        # Decode the response
        response = tokenizer.decode(response_ids[:, input_ids.shape[-1]:][0], skip_special_tokens=True)
        
        return {"response": response}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/api/health")
async def health_check():
    return {"status": "healthy"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run("main:app", host="127.0.0.1", port=8000, reload=True) 