J.A.R.V.I.S ~ Deep Learning/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
torch
transformers==4.38.2
numpy==1.26.4
pydantic==2.6.3
python-multipart==0.0.9 